requirements.txt
# LangChain core stack (0.0.x legacy, Windows-safe)
langchain==0.3.20
langchain-chroma==0.2.2
langchain-community==0.3.18
langchain-core==0.3.45
langchain-openai==0.3.7
langsmith==0.1.147

# Vector DB
chromadb==0.6.3
langchain-chroma==0.2.2

# Text splitters
langchain-text-splitters==0.3.6

# OpenAI
openai>=1.6.0,<2.0.0
tiktoken>=0.5.2

# Loaders / utils
pypdf>=3.17.0
python-dotenv==1.0.1

# Pydantic v1 only
pydantic==2.10.6

# Opentelemetry
opentelemetry-api==1.31.0
opentelemetry-exporter-otlp-proto-common==1.31.0
opentelemetry-exporter-otlp-proto-grpc==1.31.0
opentelemetry-exporter-otlp-proto-http==1.31.0
opentelemetry-instrumentation==0.52b0
opentelemetry-instrumentation-asgi==0.52b0
opentelemetry-instrumentation-fastapi==0.52b0
opentelemetry-proto==1.31.0
opentelemetry-sdk==1.31.0
opentelemetry-semantic-conventions==0.52b0
opentelemetry-util-http==0.52b0

# Windows + protobuf / telemetry 안정화
protobuf==5.29.3
grpcio-status==1.71.0rc2

# Streamlit
streamlit==1.42.2
streamlit-extras==0.5.5

# pysqlite3-binary
pysqlite3-binary==0.5.3; platform_system == "Linux"
